add LocationsView
